# backend/src/main/java/com/devloom/ai/toolbox/auth/dto/request/RefreshTokenRequest.java
package com.devloom.ai.toolbox.auth.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class RefreshTokenRequest {

    @NotBlank
    private String refreshToken;
}