# backend/src/main/java/com/devloom/ai/toolbox/auth/dto/request/RegisterRequest.java
package com.devloom.ai.toolbox.auth.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class RegisterRequest {

    @Email
    @NotBlank
    private String email;

    @NotBlank
    @Size(min = 8, max = 64)
    private String password;

    @NotBlank
    private String code;

    @NotBlank
    private String nickname;

    private String device;
}