# backend/src/main/java/com/devloom/ai/toolbox/auth/dto/response/TokenResponse.java
package com.devloom.ai.toolbox.auth.dto.response;

import lombok.AllArgsConstructor;
import lombok.Getter;

@Getter
@AllArgsConstructor
public class TokenResponse {
    private final String accessToken;
    private final String refreshToken;
    private final long expiresIn;
}