# docs/design/auth/openapi.yml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: >
    登录、注册、鉴权及第三方接入接口定义。响应结构遵循 docs/specs/api.md，
    所有接口返回 {code,message,data} 的 JSON 格式。
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
tags:
  - name: Auth
    description: 用户注册 / 登录 / 安全操作
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 用户注册（邮箱）
      description: 提交邮箱验证码及密码创建账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRegister"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: 邮箱已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      description: 根据邮箱/手机号/第三方 ID 进行密码认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/token/refresh:
    post:
      tags: [Auth]
      summary: 刷新 Access Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 登出当前设备或所有设备
      description: 登出并吊销 Refresh Token，可选择只登出当前设备或登出所有设备
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/code/email:
    post:
      tags: [Auth]
      summary: 请求邮箱验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCodeRequest"
      responses:
        "200":
          description: 发送成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "429":
          description: 触发频率限制
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  /api/v1/auth/password/reset:
    post:
      tags: [Auth]
      summary: 重置密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: 重置成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/oauth/linuxdo/authorize:
    get:
      tags: [Auth]
      summary: LinuxDo OAuth 授权
      description: 重定向到 LinuxDo OAuth 授权页面，开始登录流程
      parameters:
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
          description: 登录成功后的前端跳转地址
      responses:
        "302":
          description: 重定向到 LinuxDo OAuth 授权页面
  /api/v1/auth/oauth/linuxdo/callback:
    get:
      tags: [Auth]
      summary: LinuxDo OAuth 回调
      description: 处理 LinuxDo 授权回调，完成登录或注册
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: LinuxDo 返回的授权码
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: 防 CSRF 的 state 参数
      responses:
        "200":
          description: 登录/注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseToken"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/bind/linuxdo:
    post:
      tags: [Auth]
      summary: 绑定 LinuxDo 账号
      description: 将当前已登录用户与 LinuxDo 账号进行绑定
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindLinuxDoRequest"
      responses:
        "200":
          description: 绑定成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: LinuxDo 账号已被其他用户绑定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  /api/v1/auth/profile:
    get:
      tags: [Auth]
      summary: 查询个人资料
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误或请求非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
    Unauthorized:
      description: 未认证或 Token 已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
    ApiResponseError:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
      example:
        code: 2004
        message: password error
        data: null
    ApiResponseRegister:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RegisterResponse"
    ApiResponseToken:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TokenResponse"
    ApiResponseProfile:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProfileResponse"
    RegisterRequest:
      type: object
      required: [email, password, code, nickname]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
        code:
          type: string
          example: "123456"
        nickname:
          type: string
          example: DevUser
    RegisterResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: |
            登录标识符：
            - 当前支持：邮箱地址（格式：user@example.com）
            - 未来可能支持：手机号、用户名等
          example: "user@example.com"
        type:
          type: string
          enum: [EMAIL]
          default: EMAIL
          description: |
            登录类型（可选参数）：
            - EMAIL: 邮箱登录

            注意：
            - 若不传此参数，后端会根据 identifier 格式自动识别
            - 当前版本仅支持邮箱登录，未来可能扩展支持手机号等
        password:
          type: string
          format: password
          minLength: 8
          description: 登录密码
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          description: Access Token 有效期（秒）
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Refresh Token
    LogoutRequest:
      type: object
      properties:
        scope:
          type: string
          enum: [CURRENT, ALL]
          default: CURRENT
          description: |
            登出范围：
            - CURRENT: 只登出当前设备（默认）
            - ALL: 登出所有设备
    EmailCodeRequest:
      type: object
      required: [email, scene]
      properties:
        email:
          type: string
          format: email
          description: 邮箱地址
        scene:
          type: string
          enum: [REGISTER, RESET_PASSWORD]
          description: |
            验证码使用场景：
            - REGISTER: 用户注册
            - RESET_PASSWORD: 重置密码
    PasswordResetRequest:
      type: object
      required: [email, code, newPassword, confirmPassword]
      properties:
        email:
          type: string
          format: email
          description: 邮箱地址
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^\d{6}$'
          description: 邮箱验证码（通过 /api/v1/auth/code/email 获取，场景为 RESET_PASSWORD）
        newPassword:
          type: string
          format: password
          minLength: 8
          description: 新密码
        confirmPassword:
          type: string
          format: password
          minLength: 8
          description: 确认新密码（必须与 newPassword 一致）
    BindLinuxDoRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
          description: LinuxDo OAuth 返回的授权码
        state:
          type: string
          description: 防 CSRF 的 state 参数
    ProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        avatar:
          type: string
        status:
          type: integer
          example: 1
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/AuthBindingResponse"
    AuthBindingResponse:
      type: object
      properties:
        type:
          type: string
          enum: [EMAIL, LINUX_DO]
          description: 绑定类型
        identifier:
          type: string
          description: 邮箱地址或 LinuxDo 用户名
        verified:
          type: boolean
          description: |
            是否已验证：
            - EMAIL: 通过验证码验证为 true
            - LINUX_DO: 始终为 true